# cache: stop always evaluating debug logging arguments in hotpaths