Arguments to log functions still get evaluated by Go even though
the Log() function will never be called due to the V(5).

So stop burning a ton of time constructing debug logging strings
that will never get printed anywhere, unless we've actually
got debug logging enabled.

```
Before
---
BenchmarkPopulate2UpdateArray-8   	      43	  25525739 ns/op
BenchmarkPopulate2UpdateArray-8   	      42	  24999238 ns/op
BenchmarkPopulate2UpdateArray-8   	      46	  24176152 ns/op

BenchmarkPopulate2UpdateArray-8   	      43	  26409268 ns/op
BenchmarkPopulate2UpdateArray-8   	      46	  24408124 ns/op
BenchmarkPopulate2UpdateArray-8   	      43	  25397653 ns/op

BenchmarkPopulate2UpdateArray-8   	      43	  25063454 ns/op
BenchmarkPopulate2UpdateArray-8   	      44	  24937264 ns/op
BenchmarkPopulate2UpdateArray-8   	      46	  24411459 ns/op

After
---
BenchmarkPopulate2UpdateArray-8   	      51	  22622261 ns/op
BenchmarkPopulate2UpdateArray-8   	      54	  21740579 ns/op
BenchmarkPopulate2UpdateArray-8   	      51	  22653041 ns/op

BenchmarkPopulate2UpdateArray-8   	      50	  21653251 ns/op
BenchmarkPopulate2UpdateArray-8   	      54	  22234639 ns/op
BenchmarkPopulate2UpdateArray-8   	      52	  22120362 ns/op

BenchmarkPopulate2UpdateArray-8   	      54	  22482835 ns/op
BenchmarkPopulate2UpdateArray-8   	      52	  21854665 ns/op
BenchmarkPopulate2UpdateArray-8   	      48	  22153154 ns/op
```